Sprint 3
Atualização no upload de imagens, atualização na criação de pedidos, atualização no put pedidos.

+++ src/produtos/produtos.service.ts
@@ -3,17 +3,18 @@ import { CreateProdutoDto } from './dto/create-produto.dto';
 import { UpdateProdutoDto } from './dto/update-produto.dto';
 import { PrismaService } from 'src/prisma.service';
 import { Produto } from './entities/produto.entity';
+import { UploadService } from '../upload/upload.service';
 
 @Injectable()
 export class ProdutosService {
-  constructor(private prisma: PrismaService) { }
+  constructor(private prisma: PrismaService, private upload: UploadService) { }
   async create(createProdutoDto: CreateProdutoDto) {
 
     const data = { ...createProdutoDto }
 
     return await this.prisma.produtos.create({
       data:
-        { descricao: data.descricao, quantidade_estoque: data.quantidade_estoque, valor: data.valor, categoria_id: data.categoria_id, produto_imagem: data.imagem || null }
+        { descricao: data.descricao, quantidade_estoque: data.quantidade_estoque, valor: data.valor, categoria_id: data.categoria_id, produto_imagem: data.produto_imagem || null }
 
     });
 
@@ -41,8 +42,21 @@ export class ProdutosService {
       throw new NotFoundException(`Produto com id ${id} não encontrado`)
     }
 
+
+
     try {
-      await this.prisma.produtos.update({ where: { id: id }, data: { descricao: data.descricao, quantidade_estoque: data.quantidade_estoque, valor: data.valor, categoria_id: data.categoria_id } })
+      if (data.produto_imagem) {
+
+        const consulta = await this.upload.consultarImagem(data.produto_imagem);
+        if (data.produto_imagem == produto.produto_imagem) {
+          throw new BadRequestException(`Imagem já cadastrada`)
+        }
+        await this.upload.deleteArquivo(produto.produto_imagem)
+
+        await this.prisma.produtos.update({ where: { id: id }, data: { descricao: data.descricao, quantidade_estoque: data.quantidade_estoque, valor: data.valor, categoria_id: data.categoria_id, produto_imagem: data.produto_imagem } })
+      } else {
+        await this.prisma.produtos.update({ where: { id: id }, data: { descricao: data.descricao, quantidade_estoque: data.quantidade_estoque, valor: data.valor, categoria_id: data.categoria_id, produto_imagem: null } })
+      }
     } catch (error) {
       throw new BadRequestException(error.message)
 
@@ -53,19 +67,28 @@ export class ProdutosService {
   }
 
   async remove(id: number) {
+    const Produto = await this.prisma.produtos.findUnique({ where: { id: id } })
+
+    if (!Produto) {
+      throw new NotFoundException(`Produto com id ${id} não encontrado`)
+    }
+    const produtoPedido = await this.prisma.pedidoProduto.findFirst({ where: { produto_id: id } })
+
+    if (produtoPedido) {
+      throw new BadRequestException(`Produto com id ${id} não pode ser excluído pois está vinculado a um pedido`)
+    }
 
     try {
       await this.prisma.produtos.delete({ where: { id: id } })
+      if (Produto.produto_imagem) {
+        await this.upload.deleteArquivo(Produto.produto_imagem)
+      }
       return { message: `Produto com id ${id} deletado com sucesso` }
     } catch (error) {
       throw new BadRequestException("Erro ao excluir produto")
     }
-    const produto = await this.prisma.produtos.findUnique({ where: { id: id } })
 
-    if (!produto) {
-      throw new NotFoundException(`Produto com id ${id} não encontrado`)
-    }
 
-    return produto
+
   }
 }
